Add files via upload
fixing bugs related to stimulus reordering:  When common stimuli are found prior to analyzing a transformatoin, they are reordered alphabetically  in psg_geomodels_run, psg_geomodels_test, and this reordering needs to be propagated to psg_majaxes.  However, for legacy compatibility, if the stimuli in the ref and adj sets are identical, then there is no reordering  This is taken care of in psg_commonstims.